pyproject.toml
[tool.poetry]
name = "pyimg"
version = "0.0.1"
description = "Aula sobre as tecnicas de machine learning (ML) e Deep Learning para analise de imagens."
authors = ["Erik Ingwersen <erik.ingwersen@br.ey.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ingwersen-erik/pyimg"
repository = "https://github.com/ingwersen-erik/pyimg"
documentation = "https://pyimg.readthedocs.io"
packages = [
    { include = "pyimg", from = "src" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/ingwersen-erik/pyimg/releases"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"

black = { version = ">=21.10b0", optional = false }
click = { version = ">=8.0.1", optional = false }
flake8 = { version = ">=4.0.1", optional = false }
flake8-annotations = {version = "^2.7.0", python = ">= 3.6.2", optional = false }
flake8-bandit = { version = ">=2.1.2", optional = false }
flake8-bugbear = { version = ">=21.9.2", optional = false }
flake8-docstrings = { version = ">=1.6.0", optional = false }
flake8-rst-docstrings = { version = ">=0.2.5", optional = false }
hydra-core = { version = "^1.1.1", optional = false }
jupyterlab = { version = "^3.3.4", optional = false }
jupyterlab-myst = { version = "^0.1.6", optional = false }
matplotlib = { version = "^3.5.1", optional = false }
omegaconf = { version = "^2.1.1", optional = false }
python-dotenv = { version = "^0.19.2", optional = false }
rich = { version = "^12.0.1", optional = false }
scikit-learn = { version = "^1.0.2", optional = false }
tensorflow = { version = "^2.8.0", optional = false }
tensorflow-addons = { version = "^0.16.1", optional = false }
nox-poetry = "^0.9.0"
nox = "^2022.1.7"
nptyping = "^2.0.0"
pytest = "^7.1.1"
user-agent = "^0.1.10"
pandas = "^1.4.2"
numpy = "^1.22.3"
Pillow = "^9.1.1"


[tool.poetry.dev-dependencies]
Pygments = { version = ">=2.10.0", optional = true }
coverage = {extras = ["toml"], version = ">=6.2" }
darglint = { version = ">=1.8.1", optional = true }
furo = { version = ">=2021.11.12", optional = true }
isort = { version = ">=5.10.1", optional = true }
mypy = { version = ">=0.930", optional = true }
myst-parser = { version = ">=0.16.1", optional = true }
pep8-naming = { version = ">=0.12.1", optional = true }
pre-commit = { version = ">=2.16.0", optional = true }
pre-commit-hooks = { version = ">=4.1.0", optional = true }
pytest = { version = ">=6.2.5", optional = true }
pyupgrade = { version = ">=2.29.1", optional = true }
safety = { version = ">=1.10.3", optional = true }
sphinx = { version = ">=4.3.2", optional = true }
sphinx-autobuild = { version = ">=2021.3.14", optional = true }
sphinx-click = { version = ">=3.0.2", optional = true }
typeguard = { version = ">=2.13.3", optional = true }
xdoctest = { extras = ["colors"], version = ">=0.15.10", optional = true }


[tool.poetry.scripts]
pyimg = "pyimg.__main__:main"


[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]


[tool.coverage.run]
branch = true
source = ["pyimg", "tests"]


[tool.coverage.report]
show_missing = true
fail_under = 100


[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2


[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
[tool.black]
line-length = 80
skip-string-normalization = true
target-version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | docs
)/
'''


[build-system]
requires = [
    "poetry-core>=1.0.0",
    "setuptools >= 40.9.0",
    "wheel",
]
build-backend = "poetry.core.masonry.api"
